Change to Int

+++ app/models.py
@@ -6,8 +6,8 @@ class DeviceLog(db.Model):
     __tablename__ = 'devicelogs'
 
     log_id = db.Column(db.Integer, primary_key=True)
-    power = db.Column(db.Float())
-    temp = db.Column(db.Float())
+    power = db.Column(db.Integer())
+    temp = db.Column(db.Integer())
     created_on = db.Column(db.DateTime, server_default=db.func.now())
     updated_on = db.Column(db.DateTime, server_default=db.func.now(), onupdate=db.func.now())
